# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Replay parser and stats API for Command \u0026 Conquer Generals / Zero Hour",
        "title": "CNC Stats API",
        "contact": {},
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/player-money": {
            "get": {
                "description": "Retrieve player money data with optional pagination.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "player-money"
                ],
                "summary": "List player money data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Maximum number of records to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of records to skip",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.PlayerMoneyListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Store a new player money data snapshot.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "player-money"
                ],
                "summary": "Create player money data",
                "parameters": [
                    {
                        "description": "Player money data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.MoneyDataRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/player-money/{id}": {
            "get": {
                "description": "Retrieve a single player money data record by its ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "player-money"
                ],
                "summary": "Get player money data by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Record ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/database.PlayerMoneyData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a single player money data record by its ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "player-money"
                ],
                "summary": "Delete player money data by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Record ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/replay": {
            "post": {
                "description": "Upload a .rep replay file and receive parsed replay data. Use format=v2 to get enhanced stats if a matching stats file exists.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "replay"
                ],
                "summary": "Parse a replay file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Replay file to parse",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "enum": [
                            "v2"
                        ],
                        "type": "string",
                        "description": "Response format",
                        "name": "format",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/zhreplay.EnhancedReplayV2"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/stats": {
            "post": {
                "description": "Receive gzip-compressed JSON stats from a Generals game and store them keyed by seed.",
                "consumes": [
                    "application/octet-stream"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "Upload game stats",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Game seed identifier",
                        "name": "X-Game-Seed",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.StatsUploadResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "database.MoneyDataRequest": {
            "type": "object",
            "properties": {
                "bombardment": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "buildings_built": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "buildings_killed": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                },
                "buildings_lost": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "faction_buildings_captured": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "generals_points_total": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "generals_points_used": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "hold_the_line": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "money": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "money_earned": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "power_total": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "power_used": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "radars_built": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "search_and_destroy": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "seed": {
                    "type": "string"
                },
                "tech_buildings_captured": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "timecode": {
                    "type": "integer"
                },
                "units_built": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "units_killed": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                },
                "units_lost": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "xp": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "xp_level": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "main.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "string",
                    "example": "underlying error message"
                },
                "error": {
                    "type": "string",
                    "example": "Something went wrong"
                }
            }
        },
        "main.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Player money data deleted successfully"
                }
            }
        },
        "main.PlayerMoneyListResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer",
                    "example": 5
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/database.PlayerMoneyData"
                    }
                }
            }
        },
        "main.StatsUploadResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Stats stored successfully"
                },
                "seed": {
                    "type": "string",
                    "example": "12345"
                },
                "size": {
                    "type": "integer",
                    "example": 8192
                }
            }
        },
...
        "zhreplay.EnhancedReplayV2": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/body.BodyChunk"
                    }
                },
                "gameInfo": {
                    "$ref": "#/definitions/zhreplay.GameInfoV2"
                },
                "header": {
                    "$ref": "#/definitions/header.GeneralsHeader"
                },
                "offset": {
                    "type": "integer"
                },
                "stats": {
                    "$ref": "#/definitions/zhreplay.EnrichedStats"
                },
                "summary": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/zhreplay.PlayerSummaryV2"
                    }
                },
                "version": {
                    "type": "integer"
                },
                "winMethod": {
                    "type": "string"
                }
            }
        },
...
    }
}